tests/socks5: Fix -Wshorten-64-to-32 warning
Clean this up (clang specific warning):

  src/gwproxy/tests/socks5.c:386:9: error: implicit conversion loses \
  integer precision: 'size_t' (aka 'unsigned long') to 'int' [-Werror,-Wshorten-64-to-32]
    386 |                 ret = (size_t)fwr;
        |                     ~ ^~~~~~~~~~~
  1 error generated.

+++ src/gwproxy/tests/socks5.c
@@ -383,7 +383,7 @@ static ssize_t file_put_contents(const char *path, const char *data, size_t len)
 	if (fwr < len)
 		ret = -EIO;
 	else
-		ret = (size_t)fwr;
+		ret = (int)fwr;
 
 	fclose(f);
 	return ret;
